removing unused code and moving Counter up
