+++ src/Nethermind/Nethermind.Network/P2P/Subprotocols/Eth/V63/Eth63ProtocolHandler.cs
@@ -31,18 +31,6 @@ using Nethermind.TxPool;
 
 namespace Nethermind.Network.P2P.Subprotocols.Eth.V63
 {
-    public class Eth64ProtocolHandler : Eth63ProtocolHandler
-    {
-        public Eth64ProtocolHandler(ISession session,
-            IMessageSerializationService serializer,
-            INodeStatsManager nodeStatsManager,
-            ISyncServer syncServer,
-            ITxPool txPool,
-            ILogManager logManager) : base(session, serializer, nodeStatsManager, syncServer, txPool, logManager)
-        {
-        }
-    }
-    
     public class Eth63ProtocolHandler : Eth62ProtocolHandler
     {
         private readonly BlockingCollection<Request<GetNodeDataMessage, byte[][]>> _nodeDataRequests
@@ -67,36 +55,24 @@ namespace Nethermind.Network.P2P.Subprotocols.Eth.V63
         public override void HandleMessage(ZeroPacket message)
         {
             base.HandleMessage(message);
-
             int size = message.Content.ReadableBytes;
+
             switch (message.PacketType)
             {
                 case Eth63MessageCode.GetReceipts:
                     if (NetworkDiagTracer.IsEnabled) NetworkDiagTracer.ReportIncomingMessage(Session.Node.Host, Name, nameof(GetReceiptsMessage));
-                    Interlocked.Increment(ref Counter);
-                    if (Logger.IsTrace) Logger.Trace($"{Counter:D5} GetReceipts from {Node:c}");
-                    Metrics.Eth63GetReceiptsReceived++;
                     Handle(Deserialize<GetReceiptsMessage>(message.Content));
                     break;
                 case Eth63MessageCode.Receipts:
                     if (NetworkDiagTracer.IsEnabled) NetworkDiagTracer.ReportIncomingMessage(Session.Node.Host, Name, nameof(ReceiptsMessage));
-                    Interlocked.Increment(ref Counter);
-                    if (Logger.IsTrace) Logger.Trace($"{Counter:D5} Receipts from {Node:c}");
-                    Metrics.Eth63ReceiptsReceived++;
                     Handle(Deserialize<ReceiptsMessage>(message.Content), size);
                     break;
                 case Eth63MessageCode.GetNodeData:
                     if (NetworkDiagTracer.IsEnabled) NetworkDiagTracer.ReportIncomingMessage(Session.Node.Host, Name, nameof(GetNodeDataMessage));
-                    Interlocked.Increment(ref Counter);
-                    if (Logger.IsTrace) Logger.Trace($"{Counter:D5} GetNodeData from {Node:c}");
-                    Metrics.Eth63GetNodeDataReceived++;
                     Handle(Deserialize<GetNodeDataMessage>(message.Content));
                     break;
                 case Eth63MessageCode.NodeData:
                     if (NetworkDiagTracer.IsEnabled) NetworkDiagTracer.ReportIncomingMessage(Session.Node.Host, Name, nameof(NodeDataMessage));
-                    Interlocked.Increment(ref Counter);
-                    if (Logger.IsTrace) Logger.Trace($"{Counter:D5} NodeData from {Node:c}");
-                    Metrics.Eth63NodeDataReceived++;
                     Handle(Deserialize<NodeDataMessage>(message.Content), size);
                     break;
             }
@@ -106,14 +82,14 @@ namespace Nethermind.Network.P2P.Subprotocols.Eth.V63
 
         private void Handle(GetReceiptsMessage msg)
         {
+            Metrics.Eth63GetReceiptsReceived++;
             if (msg.BlockHashes.Count > 512)
             {
                 throw new EthSynchronizationException("Incoming receipts request for more than 512 blocks");
             }
-            
+
             Stopwatch stopwatch = Stopwatch.StartNew();
             TxReceipt[][] txReceipts = SyncServer.GetReceipts(msg.BlockHashes);
-            Interlocked.Increment(ref Counter);
             Send(new ReceiptsMessage(txReceipts));
             stopwatch.Stop();
             if (Logger.IsTrace) Logger.Trace($"OUT {Counter:D5} Receipts to {Node:c} in {stopwatch.Elapsed.TotalMilliseconds}ms");
@@ -121,16 +97,15 @@ namespace Nethermind.Network.P2P.Subprotocols.Eth.V63
 
         private void Handle(ReceiptsMessage msg, long size)
         {
+            Metrics.Eth63ReceiptsReceived++;
             Request<GetReceiptsMessage, TxReceipt[][]> request = _receiptsRequests.Take();
-            if (IsRequestMatched(request, msg))
-            {
-                request.ResponseSize = size;
-                request.CompletionSource.SetResult(msg.TxReceipts);
-            }
+            request.ResponseSize = size;
+            request.CompletionSource.SetResult(msg.TxReceipts);
         }
 
         private void Handle(GetNodeDataMessage msg)
         {
+            Metrics.Eth63GetNodeDataReceived++;
             if (msg.Keys.Count > 4096)
             {
                 throw new EthSynchronizationException("Incoming node data request for more than 4096 nodes");
@@ -138,7 +113,6 @@ namespace Nethermind.Network.P2P.Subprotocols.Eth.V63
 
             Stopwatch stopwatch = Stopwatch.StartNew();
             byte[][] nodeData = SyncServer.GetNodeData(msg.Keys);
-            Interlocked.Increment(ref Counter);
             Send(new NodeDataMessage(nodeData));
             stopwatch.Stop();
             if (Logger.IsTrace) Logger.Trace($"OUT {Counter:D5} NodeData to {Node:c} in {stopwatch.Elapsed.TotalMilliseconds}ms");
@@ -146,12 +120,10 @@ namespace Nethermind.Network.P2P.Subprotocols.Eth.V63
 
         private void Handle(NodeDataMessage msg, int size)
         {
+            Metrics.Eth63NodeDataReceived++;
             Request<GetNodeDataMessage, byte[][]> request = _nodeDataRequests.Take();
             request.ResponseSize = size;
-            if (IsRequestMatched(request, msg))
-            {
-                request.CompletionSource.SetResult(msg.Data);
-            }
+            request.CompletionSource.SetResult(msg.Data);
         }
 
         public override async Task<byte[][]> GetNodeData(IList<Keccak> keys, CancellationToken token)
@@ -173,12 +145,8 @@ namespace Nethermind.Network.P2P.Subprotocols.Eth.V63
                 return Array.Empty<TxReceipt[]>();
             }
 
-            // Logger.Info($"Sending receipts request ({blockHashes.Count}) to {this}");
-
             GetReceiptsMessage msg = new GetReceiptsMessage(blockHashes);
             TxReceipt[][] txReceipts = await SendRequest(msg, token);
-
-            // Logger.Info($"Sent receipts request ({blockHashes.Count}) to {this} - received {txReceipts.Length}");
             return txReceipts;
         }
 
@@ -259,22 +227,6 @@ namespace Nethermind.Network.P2P.Subprotocols.Eth.V63
             StatsManager.ReportTransferSpeedEvent(Session.Node, 0L);
 
             throw new TimeoutException($"{Session} Request timeout in {nameof(GetReceiptsMessage)}");
-        }
-
-        [Todo(Improve.MissingFunctionality, "Need to compare response")]
-        private bool IsRequestMatched(
-            Request<GetNodeDataMessage, byte[][]> request,
-            NodeDataMessage response)
-        {
-            return response.PacketType == Eth63MessageCode.NodeData; // TODO: more detailed
-        }
-
-        [Todo(Improve.MissingFunctionality, "Need to compare response")]
-        private bool IsRequestMatched(
-            Request<GetReceiptsMessage, TxReceipt[][]> request,
-            ReceiptsMessage response)
-        {
-            return response.PacketType == Eth63MessageCode.Receipts; // TODO: more detailed
         }
     }
 }
